requirements.txt
aiogram==3.22.0
python-dotenv==1.1.1
pydantic==2.11.7
pydantic-settings==2.11.0

SQLAlchemy==2.0.44
asyncpg==0.30.0
alembic==1.17.0
psycopg==3.2.11
psycopg-binary==3.2.11

celery==5.5.3
redis==7.0.1

yookassa==3.7.1
requests==2.32.5

Pillow==12.0.0

aiohttp==3.12.15
aiofiles==24.1.0

coloredlogs==15.0.1


typing_extensions==4.15.0